+++ flake.nix
@@ -32,7 +32,7 @@
 
         sentinel = pkgs.rustPlatform.buildRustPackage {
           pname = "sentinel";
-          version = "0.6.0";
+          version = "0.6.1";
           src = ./.;
 
           cargoLock = {